+++ src/Keymaps/asdf_keymap_apple2_caps.c
@@ -31,6 +31,7 @@
 #include "asdf_virtual.h"
 #include "asdf_modifiers.h"
 #include "asdf_keymap_apple2_add_map.h"
+#include "asdf_keymap_apple2.h"
 #include "asdf_keymap_apple2_caps.h"
 #include "asdf_keymap_apple2_maps.h"