+++ src/Keymaps/asdf_keymap_classic.h
@@ -27,6 +27,9 @@
 #if !defined(ASDF_KEYMAP_DEFS_CLASSIC_H)
 #define ASDF_KEYMAP_DEFS_CLASSIC_H
 
+#include "asdf_arch.h"
+#include "asdf_keymaps.h"
+
 
 // The size of this keymap's matrices, at most ASDF_MAX_ROWS by ASDF_MAX_COLS.
 // Keys not given in the YAML matrices do nothing.
@@ -52,6 +55,9 @@
 
 
 #define ASDF_CLASSIC_PRINT_SPEED 40
+
+/** The "classic" ADM 3A style keymap's descriptor, in flash. */
+extern const asdf_keymap_t FLASH classic_keymap;
 
 #endif /* !defined (ASDF_KEYMAP_DEFS_CLASSIC_H) */