+++ src/Keymaps/asdf_keymap_classic_caps.h
@@ -1,0 +1,37 @@
+// -*- mode: C; tab-width: 4 ; indent-tabs-mode: nil -*-
+/**
+ * @file asdf_keymap_classic_caps.h
+ *
+ * The ALL CAPS "classic" keymap's descriptor. It shares its matrix size and
+ * outputs with the "classic" keymap (asdf_keymap_classic.h).
+ *
+ * Part of the Unified Keyboard Project ASDF keyboard firmware.
+ *
+ * @copyright Copyright 2019 David Fenyes. GNU General Public License
+ * version 3 or later; see the license notice below.
+ */
+//
+// This program is free software: you can redistribute it and/or modify it under
+// the terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+// details.
+//
+// You should have received a copy of the GNU General Public License along with
+// this program. If not, see <https://www.gnu.org/licenses/>.
+
+
+#if !defined(ASDF_KEYMAP_DEFS_CLASSIC_CAPS_H)
+#define ASDF_KEYMAP_DEFS_CLASSIC_CAPS_H
+
+#include "asdf_arch.h"
+#include "asdf_keymaps.h"
+
+/** The ALL CAPS "classic" keymap's descriptor, in flash. */
+extern const asdf_keymap_t FLASH classic_caps_keymap;
+
+#endif /* !defined (ASDF_KEYMAP_DEFS_CLASSIC_CAPS_H) */