+++ src/asdf_print.c
@@ -45,8 +45,8 @@ void asdf_print_flash(asdf_t *kb, const char *str)
 {
   char c;
 
-  while ((c = (char) FLASH_READ(str++))) {
-    asdf_putc(kb, c);
+  while ((c = (char) FLASH_READ(str++)) != '\0') {
+    (void)asdf_putc(kb, c);
   }
 }